Remade sync for testing

+++ systems/SyncPos2DSystem.cpp
@@ -34,7 +34,7 @@ void		SyncPos2DSystem::processEntity(Entity *entity, const float)
       pos_comp->setY(sync_comp->getSyncPos().getY());
       return ;
     }
-  ratio = network_comp->getLastUpdate() / network_comp->getPreviousUpdateTime();
+  ratio = static_cast<float>(network_comp->getLastUpdate()) / network_comp->getPreviousUpdateTime();
   delta_x = (sync_comp->getSyncPos().getX() - pos_comp->getX()) * ratio;
   delta_y = (sync_comp->getSyncPos().getY() - pos_comp->getY()) * ratio;
   pos_comp->setX(pos_comp->getX() + delta_x);
